apps/ms_proposals/requirements.txt
Faker==33.1.0
psycopg2-binary==2.9.6